Add Import statements for pictures and two test pictures
These import statements are neecessary for webpack to recognize the
images as dependencies and pack them into the bundle (see [1] for
details).

We can set the directory property of vue-photo-gallery to '.'
since the image names contain the whole path of the images.

In the former implementation, the image paths were generated at runtime,
so webpack did not have a chance to include the images.

[1] https://medium.com/a-beginners-guide-for-webpack-2/handling-images-e1a2a2c28f8d